Reporter transactions filters for date and void
The original transactions listing listed every transaction, now that
void has been implemented in the API the default should exclude
transactions that are tagged "Void". The user should be encouraged to
void transactions (rather than losing information in a delete) and the
features should model void transactions as seamlessly as deleting.

Also included is the skeleton for adding a date filter for the query.
Defaulting to 100 years before and after current date. Will need to
implement the command line options eventually.