Get Listing RPC Message implemented
The RPC endpoints are not the ideal place for queries to the GL however
can foresee them possibly being needed. The GetListing query would be
used to return all transactions between the provided dates. Its rough
and slow but if faster and more detail is needed the user should be
encouraged to query the database directly