Add Decimal Places to Ledger_CLI and Reporter
The amounts are stored in the databases as integers, originally was
planning on keeping everything in the ledger_cli and reporter
executables as integers aswell and enforcing the use of smallest units
as integers (cents) rather than allowing for the larger units (dollars).

However the code pulled from https://github.com/howeyc/ledger for the
parsing of ledger text files was processing decimals and the
datastructure being used in ledger_cli was allowing for it. Rather than
keeping inconsistencies with some parts being integer and some parts
allowing decimal this makes everything decimal.

This will cause a breaking change as inputs will now be for whole
dollars and will then save into the database a cents. Previously if a
person had been using the api as cents they will need to update it for
dollars. IF a person had been using the api as dollars previously then
all of their data will be stored as a number off by a factor of 100.