GH-84 initial sample dockerfile and helper make targets
allows running 'make build-docker' (or 'make docker-build')
to build the 3 godbledger binaries into a docker image.

'make docker-start' will start it up on docker forwarding
port 50051

'make docker-stop' will stop the container

'make docker-clean' will remove the container from local
docker engine

'make docker-login' will log into an ash shell in the
container, mounting it if it's not already running, or
exec'ing into the running container if it is.

this is just a first draft, and currently when I run
'ledger_cli single' on my host OS it seems to connect
to godbledger-server running inside Docker but then
fails. maybe I'm not using the command correctly?