GH-84 fix rpc host (inside docker must listen wider than localhost)
godbledger server in the docker container cannot listen
only to 'localhost' as 'localhost' excludes traffic from
outside the container.

instead we tell godbledger server to listen for rpc traffic
from any host (by default) allowing it to receive traffic
from the host machine and leaving it up to the operator
to secure traffic into the server as needed by the host
environment.