GH-84 mount host ledgerdata folder for use by godbledger-server container
this is a bit tricky, as it's making a number of assumptions, but the basic
idea is that we want the godbledger instance running in docker to share
the data directory on the host machine.

this allows the developer running 'make docker-start' to send transactions
to localhost:50051 (for example) which get routed into the godbledger-server
container and handled by the godbledger app running in Docker, which is
configured to write to the same data directory as the host machine, so that
transactions can be reported on by the reporter CLI tool (which bypasses
the godbledger-server and reads the database directly).

the datadirectory used by the makefile when it starts the godbledger-server
container can be customized by overriding GDBL_DATA_DIR

    GDBL_DATA_DIR=$(HOME)/.ledger-dir make docker-start