GH-84 pick a default data_dir based on os
this should allow 'make docker-start' to set up the
same default data directory used by the app code

in the future we might make this coupling (between
the docker commands in the makefile and the startup
code in the apps) more explicit by updating the
makefile to shell out and invoke some of the go
code to ask for the default data directory, allowing
the same logic to drive both choices.

for now, however, this seems to work, allowing the
default data directory on MacOS and ubuntu and
Windows to all map properly.